+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export { default as GameSelect } from './GameSelect';
 export { default as SpawnTypeFilter } from './SpawnTypeFilter';
 export { default as VoiceCommandButton } from './VoiceCommandButton';
+export { default as ToggleTipButton } from './ToggleTipButton';